requirements.txt
matplotlib==3.9.2
numpy==2.1.0
Pillow==10.4.0
pyModbusTCP==0.2.2
Requests==2.32.3
SQLAlchemy==2.0.32
keyboard==0.13.5
pyserial==3.5
python-dotenv>=1.0.0